Fix to check for method getId

+++ Listener/DeleteListener.php
@@ -16,12 +16,14 @@ class DeleteListener {
         $entity = $args->getEntity();
         $entityManager = $args->getEntityManager();
         
-
-        $query = $entityManager->createQuery("select a from MaithCommonAdminBundle:mAlbum a where a.object_id = :id and a.object_class = :object_class")->setParameters(array('id' => $entity->getId(), 'object_class' => get_class($entity)));
-        $albums = $query->getResult();
-        foreach($albums as $album)
+        if(method_exists($entity, 'getId'))
         {
-            $entityManager->remove($album);
+            $query = $entityManager->createQuery("select a from MaithCommonAdminBundle:mAlbum a where a.object_id = :id and a.object_class = :object_class")->setParameters(array('id' => $entity->getId(), 'object_class' => get_class($entity)));
+            $albums = $query->getResult();
+            foreach($albums as $album)
+            {
+                $entityManager->remove($album);
+            }    
         }
     }
 }
